Implement ReplayableHand
This class maps a raw hand to N (matrix, street) tuples.
N is equal to the total number of actions taken across all streets.
`street` is an int to assist with round based filtering

I also added an `updated_parser` notebook.
This notebook shows how to use the parser to map a raw hand to matrices.
It can be used to verify the integrity of the mapping implementation.